Sifat

+++ src/components/EventCard/EventCard.jsx
@@ -13,6 +13,7 @@ import {
   FaShare,
   FaTrash,
 } from "react-icons/fa6";
+import Swal from "sweetalert2";
 
 const EventCard = ({ event }) => {
   const queryClient = useQueryClient();
@@ -27,7 +28,6 @@ const EventCard = ({ event }) => {
     shareableLink,
   } = event;
 
-  // console.log('event card', event)
   const meetLinkRef = useRef(null);
   const router = useRouter();
 
@@ -61,16 +61,26 @@ const EventCard = ({ event }) => {
     onSuccess: (data) => {
       console.log(data);
       if (data === "Item Deleted") {
-        // Todo toast should be shown
+        toast.success("Event successfully deleted!");
         queryClient.invalidateQueries(["singleEventDataGet"]);
       }
     },
   });
 
   const handleDelete = async (id) => {
-    // Todo a alert should be shown that the user is sure something like that
-    // calling mutateAsync function
-    deleteMutateAsync(id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteMutateAsync(id);
+      }
+    });
   };
 
   // Finish functionality---------
@@ -93,7 +103,7 @@ const EventCard = ({ event }) => {
     onSuccess: (data) => {
       if (data.status === 200) {
         // Todo implement toast
-        // toast.success('Lead successfully updated!');
+        toast.success('Lead successfully updated!');
         queryClient.invalidateQueries(["singleEventDataGet"]);
       }
     },
